requirements.txt
pandas
supabase
openpyxl
python-dotenv
streamlit
google-generativeai
